use %ld and (long) casts for time_t values in *printf
git-svn-id: https://svn.code.sf.net/p/gphoto/code/trunk/libgphoto2@7741 67ed7778-7388-44ab-90cf-0a291f65f57c

+++ camlibs/canon/library.c
@@ -804,8 +804,8 @@ camera_summary (Camera *camera, CameraText *summary, GPContext *context)
 		  (long)tmp_time, (long)local_time, (long)tm->tm_gmtoff);
 #else
 	local_time = tmp_time - timezone;
-	GP_DEBUG ("camera_summary: converted %i to localtime %i (timezone is %i)",
-		  tmp_time, local_time, timezone);
+	GP_DEBUG ("camera_summary: converted %ld to localtime %ld (timezone is %ld)",
+		  (long)tmp_time, (long)local_time, (long)timezone);
 #endif
 
 	if (res == GP_OK) {
